island: move PhoneStreamCoordinator into island (Type-B coordinator #3)
Relocates the streaming domain (RTMP/SRT/WHIP publishing) into island:

- PhoneStreamCoordinator.ts + StreamLifecycleController.ts + cloudStreamApi.ts ->
  modules/island/src/services/ (git mv). Host couplings resolved: cloudClient.{start,
  get,stop}ManagedStream -> cloudClientService (island); getRuntimeHooks().glassesStatus
  -> direct island-store read; btsdk -> internal path. Dead v2StreamApi.ts removed.
- LocalMiniappRuntime drives phoneStreamCoordinator directly (start/stop/status-fanout)
  instead of getRuntimeHooks().streaming across all 5 sites.
- Removes the streaming hook from the RuntimeHooks seam (StreamingAdapter +
  StreamPublisherStartResult + ManagedStreamStartResult deleted) and the host
  MantleManager configureRuntime binding. MantleManager's stream_status / keep_alive_ack
  listeners keep the coordinator, now imported from @mentra/island; its
  wireStreamingStatusFanout() call now subscribes the island coordinator directly.
- Exports phoneStreamCoordinator from the island index; jest mock provides it.

Both stream bun:tests moved to modules/island/src/services/__tests__/ and kept GREEN
(17 + 6 pass) — repointed mock paths + mocked the glasses store/GlassesReadiness (the
real store transitively drags react-native, which bun can't parse).

tsc + island emit build clean; full jest 282 green.
