+++ mobile/modules/island/src/index.ts
@@ -50,6 +50,7 @@ export {
 export {default as localSttFallbackCoordinator} from "./services/LocalSttFallbackCoordinator"
 export {default as micStateCoordinator} from "./services/MicStateCoordinator"
 export {phonePhotoCoordinator} from "./services/PhonePhotoCoordinator"
+export {phoneStreamCoordinator} from "./services/PhoneStreamCoordinator"
 export {
   default as sttModelManager,
   STTModelManager,
@@ -160,7 +161,6 @@ export {
   type SettingsAccessor,
   type StoreAccessor,
   type GlassesSnapshot,
-  type StreamingAdapter,
   type InteropAdapter,
   type InteropAuditEvent,
 } from "./runtime/config"
