+++ mobile/src/services/MantleManager.ts
@@ -12,7 +12,7 @@ import headingService from "@/services/HeadingService"
 import {bootstrapMentraJS} from "@/services/mentraJsBootstrap"
 import navigationService from "@/services/NavigationService"
 import {phonePhotoCoordinator} from "@mentra/island"
-import {phoneStreamCoordinator} from "@/services/streaming/PhoneStreamCoordinator"
+import {phoneStreamCoordinator} from "@mentra/island"
 import miniappCatalog from "@/services/miniapps/MiniappCatalog"
 import {preinstalledMiniappSync} from "@/services/miniapps/preinstalledMiniappSync"
 import {BUNDLED_MINIAPPS} from "@/generated/bundledMiniapps"
@@ -345,12 +345,6 @@ class MantleManager {
       locationTier: {
         setLocationTier: (rate) => this.setLocationTier(rate),
       },
-      streaming: {
-        startUnmanaged: (pkg, opts) => phoneStreamCoordinator.startUnmanaged(pkg, opts),
-        startManaged: (pkg, opts) => phoneStreamCoordinator.startManaged(pkg, opts),
-        stop: (pkg, streamId) => phoneStreamCoordinator.stop(pkg, streamId),
-        setStatusSubscriber: (cb) => phoneStreamCoordinator.setStatusSubscriber(cb),
-      },
       wifiSetup: {
         // session.glasses.requestWifiSetup → open the phone's glasses Wi-Fi
         // setup flow (same screen pairing/deeplinks use for "wifi-setup"). The
@@ -362,8 +356,11 @@ class MantleManager {
           router.push({pathname: "/wifi/scan" as any, params: (reason ? {reason} : {}) as any})
         },
       },
+      // streaming (RTMP/SRT/WHIP) moved into island (PhoneStreamCoordinator, called
+      // directly by the runtime) — no longer a host hook. (The stream_status /
+      // keep_alive_ack listeners below still route events to the coordinator.)
     })
-    // Wire the runtime's status fanout now that the streaming hook is in.
+    // Wire the runtime's status fanout (now subscribes the island coordinator directly).
     localMiniappRuntime.wireStreamingStatusFanout()
 
     // DisplayProcessor's singleton was constructed at module load — before runtime
